requirements.txt
black==24.4.2
cachetools==5.4.0
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
exceptiongroup==1.1.1
filelock==3.15.4
idna==3.7
iniconfig==2.0.0
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.11.1
platformdirs==4.2.2
pluggy==1.5.0
pyproject-api==1.7.1
pytest==8.3.2
PyYAML==6.0.1
requests==2.32.3
tomli==2.0.1
tox==4.16.0
typing_extensions==4.12.2
urllib3==1.26.19
vcrpy==6.0.1
virtualenv==20.29.2
wrapt==1.16.0
yarl==1.9.4
